tests/wpt/tests/css/css-multicol/parsing/column-span-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Multi-column Layout: parsing column-span with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-multicol/#propdef-column-span">
<meta name="assert" content="column-span supports only the grammar 'none | all'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("column-span", "auto");
test_invalid_value("column-span", "none all");
</script>
</body>
</html>
